refactor: move account-creation utility to utils directory and update imports
