+++ components/auth/oauth-github.tsx
@@ -1,4 +1,4 @@
-import createAccount from "@/util/account-creation";
+import createAccount from "@/utils/account-creation";
 import { useAuth, useSSO } from "@clerk/clerk-expo";
 import * as AuthSession from "expo-auth-session";
 import * as WebBrowser from "expo-web-browser";
